Introduced pre_init() application module callback.
pre_init is a pointer to a function with same signature as init
but executed before changing working directory and environment change.

pre_init can be NULL in this case execution is skipped.

Java module need to store absolute path to modules directory when modules
was configured as a relative path. Path to modules used in Java module to
find jar files required to start application.